chore: 포인트, 카트 redis save 코드 분리 및 테스트 코드 수정

+++ src/main/java/com/ssg/webpos/dto/PaymentsDTO.java
@@ -5,13 +5,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @ToString
 public class PaymentsDTO {
   private String name; // 상품이름
   private String pay_method; // 결제 방법
-  private int amount; // 결제 금액
+  private BigDecimal amount; // 결제 금액
   private boolean success; // 결제 성공 여부
   private String error_msg; // 실패 시 에러 메시지
   private String imp_uid;
